Use enumValue for variants
