+++ packages/snaps-cli/package.json
@@ -54,6 +54,7 @@
     "@babel/plugin-transform-runtime": "^7.13.2",
     "@babel/preset-env": "^7.23.2",
     "@babel/preset-typescript": "^7.23.2",
+    "@metamask/snaps-sdk": "workspace:^",
     "@metamask/snaps-utils": "workspace:^",
     "@metamask/snaps-webpack-plugin": "workspace:^",
     "@metamask/utils": "^8.2.1",
