+++ packages/snaps-sdk/src/internals/index.ts
@@ -1,3 +1,4 @@
 export * from './error-wrappers';
 export * from './errors';
 export * from './helpers';
+export * from './structs';
