+++ packages/snaps-utils/src/index.ts
@@ -6,7 +6,6 @@ export * from './checksum';
 export * from './deep-clone';
 export * from './default-endowments';
 export * from './entropy';
-export * from './enum';
 export * from './eval';
 export * from './errors';
 export * from './fs';
